Moving angular annotations dependencies to app/angular or old graph
