--- public/app/features/annotations/all.ts
@@ -1,5 +1,0 @@
-import { AnnotationsSrv } from './annotations_srv';
-import { eventEditor } from './event_editor';
-import { EventManager } from './event_manager';
-import { annotationTooltipDirective } from './annotation_tooltip';
-export { AnnotationsSrv, eventEditor, EventManager, annotationTooltipDirective };
