+++ public/app/features/dashboard/components/PanelEditor/getPanelFrameOptions.tsx
@@ -1,5 +1,5 @@
 import { DataLinksInlineEditor, Input, RadioButtonGroup, Select, Switch, TextArea } from '@grafana/ui';
-import { getPanelLinksVariableSuggestions } from 'app/angular/panel/panellinks/link_srv';
+import { getPanelLinksVariableSuggestions } from 'app/features/panel/panellinks/link_srv';
 import React from 'react';
 import { RepeatRowSelect } from '../RepeatRowSelect/RepeatRowSelect';
 import { OptionsPaneItemDescriptor } from './OptionsPaneItemDescriptor';
