+++ public/app/features/dashboard/components/SubMenu/DashboardLinksDashboard.tsx
@@ -2,7 +2,7 @@ import React, { useRef, useState, useLayoutEffect } from 'react';
 import { Icon, ToolbarButton, Tooltip, useStyles2 } from '@grafana/ui';
 import { sanitize, sanitizeUrl } from '@grafana/data/src/text/sanitize';
 import { getBackendSrv } from 'app/core/services/backend_srv';
-import { getLinkSrv } from '../../../../angular/panel/panellinks/link_srv';
+import { getLinkSrv } from '../../../panel/panellinks/link_srv';
 import { DashboardLink } from '../../state/DashboardModel';
 import { DashboardSearchHit } from 'app/features/search/types';
 import { selectors } from '@grafana/e2e-selectors';
