+++ public/app/features/panel/panellinks/linkSuppliers.test.ts
@@ -1,7 +1,7 @@
 import { getFieldLinksSupplier } from './linkSuppliers';
 import { applyFieldOverrides, createTheme, DataFrameView, dateTime, FieldDisplay, toDataFrame } from '@grafana/data';
 import { getLinkSrv, LinkService, LinkSrv, setLinkSrv } from './link_srv';
-import { TemplateSrv } from '../../../features/templating/template_srv';
+import { TemplateSrv } from '../../templating/template_srv';
 
 // We do not need more here and TimeSrv is hard to setup fully.
 jest.mock('app/features/dashboard/services/TimeSrv', () => ({
