+++ public/app/features/query/state/DashboardQueryRunner/AnnotationsQueryRunner.test.ts
@@ -4,7 +4,7 @@ import { AnnotationsQueryRunner } from './AnnotationsQueryRunner';
 import { AnnotationQueryRunnerOptions } from './types';
 import { silenceConsoleOutput } from '../../../../../test/core/utils/silenceConsoleOutput';
 import * as store from '../../../../store/store';
-import * as annotationsSrv from '../../../annotations/annotations_srv';
+import * as annotationsSrv from '../../../annotations/executeAnnotationQuery';
 import { Observable, of, throwError } from 'rxjs';
 import { toAsyncOfResult } from './testHelpers';
