+++ public/app/plugins/panel/graph/annotation_tooltip.ts
@@ -1,8 +1,8 @@
 import { isString, escape } from 'lodash';
 import $ from 'jquery';
 import coreModule from 'app/angular/core_module';
-import alertDef from '../alerting/state/alertDef';
-import { DashboardSrv } from '../dashboard/services/DashboardSrv';
+import alertDef from 'app/features/alerting/state/alertDef';
+import { DashboardSrv } from 'app/features/dashboard/services/DashboardSrv';
 import { ContextSrv } from 'app/core/services/context_srv';
 
 /** @ngInject */
