+++ public/app/plugins/panel/graph/event_editor.ts
@@ -2,8 +2,8 @@ import { cloneDeep, isNumber } from 'lodash';
 import { coreModule } from 'app/angular/core_module';
 import { AnnotationEvent, dateTime } from '@grafana/data';
 import { MetricsPanelCtrl } from 'app/angular/panel/metrics_panel_ctrl';
-import { deleteAnnotation, saveAnnotation, updateAnnotation } from './api';
-import { getDashboardQueryRunner } from '../query/state/DashboardQueryRunner/DashboardQueryRunner';
+import { deleteAnnotation, saveAnnotation, updateAnnotation } from '../../../features/annotations/api';
+import { getDashboardQueryRunner } from '../../../features/query/state/DashboardQueryRunner/DashboardQueryRunner';
 
 export class EventEditorCtrl {
   // @ts-ignore initialized through Angular not constructor
