+++ public/app/plugins/panel/graph/event_manager.ts
